+++ components/form-builder/app/navigation/Header.tsx
@@ -1,61 +1,16 @@
-import React, { useState, useEffect } from "react";
-import { useRouter } from "next/router";
-import LanguageToggle from "../../../globals/LanguageToggle";
-import LoginMenu from "../../../auth/LoginMenu";
+import React from "react";
 import { useSession } from "next-auth/react";
-import { useTemplateStore } from "../../store";
-
 import Link from "next/link";
 import { useAccessControl } from "@lib/hooks";
 import { useTranslation } from "next-i18next";
-import { usePublish } from "../../hooks/usePublish";
-import { useAllowPublish } from "../../hooks/useAllowPublish";
-import { Button, withMessage } from "../shared/Button";
+
+import LanguageToggle from "../../../globals/LanguageToggle";
+import LoginMenu from "../../../auth/LoginMenu";
 
 export const Header = () => {
-  const { getSchema, id, setId } = useTemplateStore((s) => ({
-    localizeField: s.localizeField,
-    getSchema: s.getSchema,
-    id: s.id,
-    setId: s.setId,
-    email: s.submission?.email,
-  }));
-
-  const [isStartPage, setIsStartPage] = useState(false);
-  const { isReady, asPath } = useRouter();
   const { status } = useSession();
-  const { isSaveable } = useAllowPublish();
   const { ability } = useAccessControl();
-
   const { t, i18n } = useTranslation(["common", "form-builder"]);
-
-  useEffect(() => {
-    if (isReady) {
-      const activePathname = new URL(asPath, location.href).pathname;
-      if (activePathname === "/form-builder") {
-        setIsStartPage(true);
-      } else {
-        setIsStartPage(false);
-      }
-    }
-  }, [asPath, isReady]);
-
-  const { uploadJson } = usePublish();
-
-  const handlePublish = async () => {
-    const schema = JSON.parse(getSchema());
-    delete schema.id;
-    delete schema.isPublished;
-
-    const result = await uploadJson(JSON.stringify(schema), id);
-
-    if (result?.id) {
-      setId(result?.id);
-    }
-  };
-
-  const ButtonWithMessage = withMessage(Button, t("saveDraftMessage", { ns: "form-builder" }));
-
   return (
     <header className="border-b-3 border-blue-dark my-10 lg:px-4 xl:px-8 px-32">
       <div className="flex justify-between">
@@ -66,11 +21,6 @@ export const Header = () => {
               {t("title", { ns: "common" })}
             </a>
           </Link>
-          {!isStartPage && isSaveable() && status === "authenticated" && (
-            <ButtonWithMessage className="ml-4" onClick={handlePublish}>
-              {t("save", { ns: "form-builder" })}
-            </ButtonWithMessage>
-          )}
         </div>
         <nav
           className="inline-flex gap-4"
