+++ components/form-builder/app/shared/index.ts
@@ -4,3 +4,4 @@ export { Dialog } from "./Dialog";
 export { Input } from "./Input";
 export { TextArea } from "./TextArea";
 export { Radio, Checkbox } from "./MultipleChoice";
+export { SaveButton } from "./SaveButton";
